Move routes into own module
